refactor(serverless): Migrate to new agent-kilt package
Ref: SMAGENT-6727
