+++ sysdig/data_source_sysdig_fargate_ECS_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/Jeffail/gabs/v2"
 	"github.com/stretchr/testify/assert"
-	"github.com/sysdiglabs/agent-kilt/runtimes/cloudformation/cfnpatcher"
+	"github.com/sysdiglabs/agent-kilt/pkg/cfnpatcher"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
